lib/py_edge_device/pyproject.toml
[tool.poetry]
name = "carlos.edge.device"
version = "0.1.1"
description = "The library for the edge device of the carlos project."
authors = ["Felix Fanghanel"]
license = "MIT"
readme = "README.md"
packages = [{include = "carlos"}]
exclude = ["**/*_test.py", "**/conftest.py", "**/test/**"]

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
loguru = "^0.7.2"
semver = "^3.0.2"
apscheduler = "^4.0.0a4"
pydantic = "^2.6.4"
pyyaml = "^6.0.1"
"carlos.edge.interface" = {path = "../py_edge_interface"}
psutil = "^5.9.8"
gpiosimulator = "^0.1"
rpi-gpio = {version = "^0.7.1", markers = "platform_machine == 'armv7l' or platform_machine == 'aarch64'"}
smbus2 = "^0.4.3"

[tool.poetry.group.dev.dependencies]
"devtools" = {path = "../py_dev_dependencies"}
types-pyyaml = "^6.0.12.20240311"
types-psutil = "^5.9.5.20240316"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
line_length = 88

[tool.mypy]
plugins = [
    "sqlalchemy.ext.mypy.plugin",
    "pydantic.mypy"
]
exclude = ['.*_test\.py$']

[tool.coverage.report]
exclude_lines = [
    # Have to re-enable the standard pragma
    "pragma: no cover",
    # Don't complain about missing debug-only code:
    "def __repr__",
    "def __hash__",
    # Don't complain if imports are not covered:
    "^import .*",
    "^from .* import .*",
    "__all__ = .*",
    # Don't complain if tests don't hit defensive assertion code:
    "raise NotImplementedError",
    # Don't complain about abstract methods, they aren't run:
    "@(abc\\.)?abstractmethod",
    # Don't complain about guard clauses
    "if __name__ == .__main__.:",
]
omit = [
    # omit all tests
    "*_test.py",
    # depends on actual device and hardware
    "carlos/edge/device/io/*.py"
]
